# Conflicts:
#	spherical_laser_projector/svg2c.py

+++ spherical_laser_projector/svg2c.py
@@ -78,7 +78,7 @@ def generate_code(paths):
 
 
 def round_float(number):
-    return int(round(number * ROUNDING_MULTIPLIER))
+    return round(number * ROUNDING_MULTIPLIER)
 
 
 def create_function_call(func_name, arguments):
